Refactored from review
